+++ src/jest.config.js
@@ -1,0 +1,4 @@
+module.exports = {
+  setupFilesAfterEnv: ["<rootDir>/setup.js"], // or .ts for TypeScript App
+  // ...other settings
+};
